requirements.txt
torch~=2.0.0
pydantic~=1.10.4
fastapi~=0.95.1
uvicorn
#for template  #new
jinja2
opencv-python~=4.7.0.72
numpy~=1.23.5
torchvision~=0.15.1
paho-mqtt~=1.6.1
numba~=0.56.4